business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetRprtShift.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetRprtShift implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+R ");
        messageString.append(command.getParameter("Rptr Shift"));
        messageString.append("\n");
        return messageString;
    }
}
